Load `posthog-js` asynchronously when telemetry is enabled.
`posthog-js` was statically imported by `TelemetryInit`, putting 280 KiB in
the eager app chunk on every page view, including the login screen.

Move the PostHog setup into its own `PostHogProvider` component and pull it
in via `loadAsync`, so the library is only fetched once telemetry settings
report it as enabled.

Measured on a production build: initial load drops from 8264 KiB to 7983 KiB
raw, 2138 KiB to 2047 KiB gzipped.

Note that `TelemetryProvider` still imports `usePostHog` from
`posthog-js/react`, whose ESM entry imports `posthog-js`, so the chunk is
still fetched alongside `LoggedInPage` for logged-in users.
